+++ src/parsers/EnglishParser.cpp
@@ -21,7 +21,6 @@ namespace lmms::gui::editor::pianoroll::parsing
     {
         this->_name = "Western / English";
         this->_identifier = "english";
-        this->_completed = true;
     };
 
     /**