+++ src/parsers/GermanParser.h
@@ -1,7 +1,7 @@
 /**
  * GermanParser.h
  *
- * Copyright (c) 2025 - 2025 Bimal Poudel <anytizer@users.noreply.github.com>
+ * Copyright (c) 2025 - 2026 Bimal Poudel <anytizer@users.noreply.github.com>
  */
 
 #ifndef LMMS_PLUGINS_MELODYEDITOR_GERMANPARSER_H