+++ HISTORY.md
@@ -7,6 +7,7 @@
  - Started to study [ALDA](https://alda.io/) and [ABC notation](https://abcnotation.com/) systems
  - Reviewing [ABC](https://abc.sourceforge.net/)
  - Using __const_cast__ instead of modifying LMMS's code, to pull piano roll in editor mode.
+ - Removed missing and none parsers.
 
 ## 2025-12